Keep logger.h internal; do not install
