+++ libtiledbsoma/src/soma/managed_query.cc
@@ -33,7 +33,7 @@
 #include "managed_query.h"
 #include <tiledb/array_experimental.h>
 #include <tiledb/attribute_experimental.h>
-#include "logger_public.h"
+#include "../utils/logger.h"
 #include "utils/common.h"
 namespace tiledbsoma {
 
@@ -218,6 +218,15 @@ std::shared_ptr<ArrayBuffers> ManagedQuery::submit_read() {
         }
     }
     return buffers_;
+}
+
+void ManagedQuery::check_column_name(const std::string& name) {
+    if (!buffers_->contains(name)) {
+        throw TileDBSOMAError(fmt::format(
+            "[ManagedQuery] Column '{}' is not available in the query "
+            "results.",
+            name));
+    }
 }
 
 };  // namespace tiledbsoma
