+++ libtiledbsoma/src/soma/soma_array.cc
@@ -32,8 +32,8 @@
 
 #include "soma_array.h"
 #include <tiledb/array_experimental.h>
+#include "../utils/logger.h"
 #include "../utils/util.h"
-#include "logger_public.h"
 namespace tiledbsoma {
 using namespace tiledb;
